Move test to more appropriate pkg
